fuck 1-1

+++ backend/src/invoices/invoices.service.ts
@@ -9,11 +9,12 @@ import { IInvoice, IInvoiceLine } from '../../../global/interfaces';
 @Injectable()
 export class InvoicesService {
     async get(user: number) {
-        return await db.models.invoice.findAll({ where: { user }, include: [db.models.invoiceLine] });
+        return await db.models.invoice.findAll({ where: { userID: user }, include: [db.models.invoiceLine] });
+        // { where: { userId: user }, include: [db.models.invoiceLine] }
     }
 
     async getID(user: number, id: number) {
-        return await db.models.invoice.findOne({ where: { user, id }, include: [db.models.invoiceLine] });
+        return await db.models.invoice.findOne({ where: { userID: user, id }, include: [db.models.invoiceLine] });
     }
 
     async post(invoice: IInvoice & { invoiceLine: IInvoiceLine[] }, userID: number) {
@@ -21,20 +22,20 @@ export class InvoicesService {
         const invoiceLine = invoice.invoiceLine;
         (<any>invoice).userID = userID;
         const invoiceDB = await db.models.invoice.create(invoice);
-        invoiceDB.setInvoices(invoiceLine);
+        invoiceDB.setInvoiceLines(invoiceLine);
         return invoiceDB;
     }
 
     async put(invoice: IInvoice, user: number) {
-        return await db.models.invoice.update(invoice, { where: { id: invoice.id!, user } });
+        return await db.models.invoice.update(invoice, { where: { id: invoice.id!, userID: user } });
     }
 
     async delete(id: number, user: number) {
-        return await db.models.invoice.destroy({ where: { id, user } });
+        return await db.models.invoice.destroy({ where: { id, userID: user } });
     }
 
     async restore(id: number, user: number) {
-        return await db.models.invoice.restore({ where: { id, user } });
+        return await db.models.invoice.restore({ where: { id, userID: user } });
     }
 
     async next(user: number) {
@@ -42,28 +43,31 @@ export class InvoicesService {
     }
 
     async check(visualID: number, user: number) {
-        return (await db.models.invoice.find({ where: { visualID, user }, attributes: ['id'] })).length !== 1;
+        return (await db.models.invoice.findAll({ where: { visualID, userID: user }, attributes: ['id'] })).length !== 1;
     }
 
     async generatePDF(id: number, user: number) {
-        const invoice = await db.models.invoice.findOne({ where: { id, user } });
+        const invoice = await db.models.invoice.findOne({ where: { id, userID: user } });
         const userDB = await db.models.user!.findOne({ where: { id: user } });
 
         if (!userDB || !invoice) throw new HttpException('No puedes generar el PDF', HttpStatus.UNAUTHORIZED);
 
         let template = fs.readFileSync(path.join(__dirname, '../shared/templates/factura.html'), 'utf-8');
+        const userValue = userDB.dataValues,
+            invoiceValues = invoice.dataValues;
+
         template = template
-            .replace(/@@nombre@@/g, userDB.name!)
-            .replace(/@@direccionUser@@/g, userDB.address!)
-            .replace(/@@telefono@@/g, userDB.phone!)
-            .replace(/@@dni@@/g, userDB.dni!)
-            .replace(/@@date@@/g, invoice.date!.toLocaleDateString())
-            .replace(/@@nombreCom@@/g, invoice.nameCompany!)
-            .replace(/@@direccion@@/g, invoice.fisicalAddress!)
-            .replace(/@@id@@/g, invoice.visualID!)
-            .replace(/@@dninie@@/g, invoice.cif!)
+            .replace(/@@nombre@@/g, userValue.name!)
+            .replace(/@@direccionUser@@/g, userValue.address!)
+            .replace(/@@telefono@@/g, userValue.phone!)
+            .replace(/@@dni@@/g, userValue.dni!)
+            .replace(/@@date@@/g, invoiceValues.date!.toLocaleDateString())
+            .replace(/@@nombreCom@@/g, invoiceValues.nameCompany!)
+            .replace(/@@direccion@@/g, invoiceValues.fisicalAddress!)
+            .replace(/@@id@@/g, `${invoiceValues.visualID!}`)
+            .replace(/@@dninie@@/g, invoiceValues.cif!)
             .replace(/@@subtotal@@/g, '------')
-            .replace(/@@notas@@/g, invoice.notes || '');
+            .replace(/@@notas@@/g, invoiceValues.notes || '');
 
         return new Promise((resolve, reject) => pdf.create(template).toBuffer((err, buf) => err ? reject(err) : resolve(buf)));
     }
